requirements.txt
langchain
streamlit
faiss-cpu
huggingface-hub
InstructorEmbedding
langchain-openai
openai
PyPDF2
sentence-transformers
langchain-community
python-dotenv